// frontend/src/app/model/User.ts
export interface User {
    id: number;
    firstname: string;
    lastname: string;
    username: string;
    password: string;
    question1: string;
    question2: string;
    question3: string;
    answer1: string;
    answer2: string;
    answer3: string;
    status: string;
    contact: string;
    role: string;
    gender: string;
    age: number;
}